Rename async() to until()
