+++ src/test/directives/until_test.ts
@@ -29,25 +29,134 @@ suite('until', () => {
     deferred = new Deferred<string>();
   });
 
-  test('displays defaultContent immediately', () => {
-    render(
-        html
-        `<div>${until(deferred.promise, html`<span>loading...</span>`)}</div>`,
-        container);
-        assert.equal(
-            stripExpressionMarkers(container.innerHTML),
-            '<div><span>loading...</span></div>');
-        deferred.resolve('foo');
-        return deferred.promise
-            .then(() => new Promise((r) => setTimeout(() => r())))
-            .then(() => {
-              assert.equal(
-                  stripExpressionMarkers(container.innerHTML),
-                  '<div>foo</div>');
-            });
+  test('renders a Promise when it resolves', async () => {
+    let resolve: (v: any) => void;
+    const promise = new Promise((res, _) => {
+      resolve = res;
+    });
+    render(html`<div>${until(promise)}</div>`, container);
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div></div>');
+    resolve!('foo');
+    await promise;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>foo</div>');
   });
 
-  test('renders new Promise over existing Promise', () => {
+  test('renders non-Promises immediately', async () => {
+    const defaultContent = html`<span>loading...</span>`;
+    render(
+        html`<div>${until(deferred.promise, defaultContent)}</div>`, container);
+    assert.equal(
+        stripExpressionMarkers(container.innerHTML),
+        '<div><span>loading...</span></div>');
+    deferred.resolve('foo');
+    await deferred.promise;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>foo</div>');
+  });
+
+  test('renders low-priority content only once', async () => {
+    const go = () => render(
+        html`<div>${until(deferred.promise, 'loading...')}</div>`, container);
+    go();
+    assert.equal(
+        stripExpressionMarkers(container.innerHTML), '<div>loading...</div>');
+    deferred.resolve('foo');
+    await deferred.promise;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>foo</div>');
+    go();
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>foo</div>');
+  });
+
+  test('renders changing defaultContent', async () => {
+    const t = (d: any) => html`<div>${until(deferred.promise, d)}</div>`;
+    render(t('A'), container);
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>A</div>');
+
+    render(t('B'), container);
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>B</div>');
+
+    deferred.resolve('C');
+    await deferred.promise;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>C</div>');
+  });
+
+  test('renders a Promise to an attribute', async () => {
+    const promise = Promise.resolve('foo');
+    render(html`<div test=${until(promise)}></div>`, container);
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div></div>');
+    await promise;
+    assert.equal(
+        stripExpressionMarkers(container.innerHTML), '<div test="foo"></div>');
+  });
+
+  test('renders defaultContent to an attribute', async () => {
+    const promise = Promise.resolve('foo');
+    render(html`<div test=${until(promise, 'bar')}></div>`, container);
+    assert.equal(
+        stripExpressionMarkers(container.innerHTML), '<div test="bar"></div>');
+    await promise;
+    assert.equal(
+        stripExpressionMarkers(container.innerHTML), '<div test="foo"></div>');
+  });
+
+  test('renders a Promise to an interpolated attribute', async () => {
+    const promise = Promise.resolve('foo');
+    render(html`<div test="value:${until(promise)}"></div>`, container);
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div></div>');
+    await promise;
+    assert.equal(
+        stripExpressionMarkers(container.innerHTML),
+        '<div test="value:foo"></div>');
+  });
+
+  test('renders defaultContent to an interpolated attribute', async () => {
+    const promise = Promise.resolve('foo');
+    render(html`<div test="value:${until(promise, 'bar')}"></div>`, container);
+    assert.equal(
+        stripExpressionMarkers(container.innerHTML),
+        '<div test="value:bar"></div>');
+    await promise;
+    assert.equal(
+        stripExpressionMarkers(container.innerHTML),
+        '<div test="value:foo"></div>');
+  });
+
+  test('renders a Promise to a property', async () => {
+    const promise = Promise.resolve('foo');
+    render(html`<div .test=${until(promise)}></div>`, container);
+    const div = container.querySelector('div');
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div></div>');
+    assert.equal((div as any).test, undefined);
+    await promise;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div></div>');
+    assert.equal((div as any).test, 'foo');
+  });
+
+  test('renders a Promise to a boolean attribute', async () => {
+    const promise = Promise.resolve(true);
+    render(html`<div ?test=${until(promise)}></div>`, container);
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div></div>');
+    await promise;
+    assert.equal(
+        stripExpressionMarkers(container.innerHTML), '<div test=""></div>');
+  });
+
+  test('renders a Promise to an event binding', async () => {
+    let called = false;
+    const promise = Promise.resolve(() => {
+      called = true;
+    });
+    render(html`<div @test=${until(promise)}></div>`, container);
+    const div = container.querySelector('div')!;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div></div>');
+    div.dispatchEvent(new CustomEvent('test'));
+    assert.isFalse(called);
+    await promise;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div></div>');
+    div.dispatchEvent(new CustomEvent('test'));
+    assert.isTrue(called);
+  });
+
+  test('renders new Promise over existing Promise', async () => {
     const t = (v: any) =>
         html`<div>${until(v, html`<span>loading...</span>`)}</div>`;
     render(t(deferred.promise), container);
@@ -62,15 +171,171 @@ suite('until', () => {
         '<div><span>loading...</span></div>');
 
     deferred2.resolve('bar');
-    return deferred2.promise.then(() => {
-      assert.equal(
-          stripExpressionMarkers(container.innerHTML), '<div>bar</div>');
+    await deferred2.promise;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>bar</div>');
 
-      deferred.resolve('foo');
-      return deferred.promise.then(() => {
-        assert.equal(
-            stripExpressionMarkers(container.innerHTML), '<div>bar</div>');
-      });
+    deferred.resolve('foo');
+    await deferred.promise;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>bar</div>');
+  });
+
+  test('renders racing Promises across renders correctly', async () => {
+    let resolve1: (v: any) => void;
+    const promise1 = new Promise((res, _) => {
+      resolve1 = res;
     });
+    let resolve2: (v: any) => void;
+    const promise2 = new Promise((res, _) => {
+      resolve2 = res;
+    });
+
+    const t = (promise: any) => html`<div>${until(promise)}</div>`;
+
+    // First render, first Promise, no value
+    render(t(promise1), container);
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div></div>');
+
+    // Second render, second Promise, still no value
+    render(t(promise2), container);
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div></div>');
+
+    // Resolve the first Promise, should not update the container
+    resolve1!('foo');
+    await promise1;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div></div>');
+    // Resolve the second Promise, should update the container
+    resolve2!('bar');
+    await promise2;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>bar</div>');
+  });
+
+  test('renders Promises resolving in high-to-low priority', async () => {
+    let resolve1: (v: any) => void;
+    const promise1 = new Promise((res, _) => {
+      resolve1 = res;
+    });
+    let resolve2: (v: any) => void;
+    const promise2 = new Promise((res, _) => {
+      resolve2 = res;
+    });
+
+    const t = () => html`<div>${until(promise1, promise2)}</div>`;
+
+    // First render with neither Promise resolved
+    render(t(), container);
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div></div>');
+
+    // Resolve the primary Promise, updates the DOM
+    resolve1!('foo');
+    await promise1;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>foo</div>');
+
+    // Resolve the secondary Promise, should not update the container
+    resolve2!('bar');
+    await promise2;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>foo</div>');
+  });
+
+  test('renders Promises resolving in low-to-high priority', async () => {
+    let resolve1: (v: any) => void;
+    const promise1 = new Promise((res, _) => {
+      resolve1 = res;
+    });
+    let resolve2: (v: any) => void;
+    const promise2 = new Promise((res, _) => {
+      resolve2 = res;
+    });
+
+    const t = () => html`<div>${until(promise1, promise2)}</div>`;
+
+    // First render with neither Promise resolved
+    render(t(), container);
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div></div>');
+
+    // Resolve the secondary Promise, updates the DOM
+    resolve2!('bar');
+    await promise2;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>bar</div>');
+
+    // Resolve the primary Promise, updates the DOM
+    resolve1!('foo');
+    await promise1;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>foo</div>');
+  });
+
+  test('renders Promises with changing priorities', async () => {
+    const promise1 = Promise.resolve('foo');
+    const promise2 = Promise.resolve('bar');
+
+    const t = (p1: any, p2: any) => html`<div>${until(p1, p2)}</div>`;
+
+    render(t(promise1, promise2), container);
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div></div>');
+    // Await a microtask to let both Promise then callbacks go
+    await 0;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>foo</div>');
+
+    render(t(promise2, promise1), container);
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>foo</div>');
+    await 0;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>bar</div>');
+  });
+
+  test('renders low-priority content when arguments change', async () => {
+    let resolve1: (v: any) => void;
+    const promise1 = new Promise((res, _) => {
+      resolve1 = res;
+    });
+    const promise2 = Promise.resolve('bar');
+
+    const t = (p1: any, p2: any) => html`<div>${until(p1, p2)}</div>`;
+
+    render(t('string', promise2), container);
+    assert.equal(
+        stripExpressionMarkers(container.innerHTML), '<div>string</div>');
+    // Await a microtask to let both Promise then callbacks go
+    await 0;
+    assert.equal(
+        stripExpressionMarkers(container.innerHTML), '<div>string</div>');
+
+    render(t(promise1, promise2), container);
+    assert.equal(
+        stripExpressionMarkers(container.innerHTML), '<div>string</div>');
+    await 0;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>bar</div>');
+    resolve1!('foo');
+    await promise1;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>foo</div>');
+  });
+
+  test('renders Promises resolving after changing priority', async () => {
+    let resolve1: (v: any) => void;
+    const promise1 = new Promise((res, _) => {
+      resolve1 = res;
+    });
+    let resolve2: (v: any) => void;
+    const promise2 = new Promise((res, _) => {
+      resolve2 = res;
+    });
+
+    const t = (p1: any, p2: any) => html`<div>${until(p1, p2)}</div>`;
+
+    // First render with neither Promise resolved
+    render(t(promise1, promise2), container);
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div></div>');
+
+    // Change priorities
+    render(t(promise2, promise1), container);
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div></div>');
+
+    // Resolve the primary Promise, updates the DOM
+    resolve1!('foo');
+    await promise1;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>foo</div>');
+
+    // Resolve the secondary Promise, also updates the DOM
+    resolve2!('bar');
+    await promise2;
+    assert.equal(stripExpressionMarkers(container.innerHTML), '<div>bar</div>');
   });
 });
