+++ test/index.html
@@ -14,7 +14,6 @@
     <script type="module" src="./lib/render_test.js"></script>
     <script type="module" src="./lib/parts_test.js"></script>
     <script type="module" src="./lib/modify-template_test.js"></script>
-    <script type="module" src="./directives/async_test.js"></script>
     <script type="module" src="./directives/async-append_test.js"></script>
     <script type="module" src="./directives/async-replace_test.js"></script>
     <script type="module" src="./directives/cache_test.js"></script>
